Testplan core: bug fix and functional improvement
* Should check the uid of real test target even if they are wrapped
  in a `Task` object as much as possible to prevent duplicate tests
  being schedule and then exception raised during creating results.
* Fix a bug that `local_runner` cannot execute a callable object.
* Fix a bug that in interactive mode a callable object that returns
  a test entity cannot be correctly added into TestRunner.
* Fix a bug that in interactive mode Multitest-parts cannot work.
* Modify uid of Multitest parts automatically to avoid conflict, and
  set it back as test entity's name for merge sibling reports.
* If fails to merge reports of MultiTest parts then error message will
  be logged and then append sibling reports at the end of plan report.
* Be cautious about the return value of step `_create_result` because
  `run` attribute in returned result of each test entity can be a
  non-boolean value.
* Fix a small issue that when unhandled exception raised during adding
  tests into plan, all tests should abort but function has no `abort`
  method, the extra error log is confusing.
* Fix a potential bug that if scheduled task return None then program
  can enter an infinite recursion.

+++ tests/functional/testplan/runners/pools/func_pool_base_tasks.py
@@ -52,17 +52,26 @@ class SuiteKillingWorker(object):
         assert env.runpath.endswith(slugify(env.cfg.name))
 
 
-def multitest_kill_one_worker(name, parent_pid, size):
+def multitest_kill_one_worker(parent_pid, size):
     """Test that kills one worker."""
     return MultiTest(
-        name="MTest{}".format(name),
-        suites=[SuiteKillingWorker(parent_pid, size)],
+        name="MTestKiller", suites=[SuiteKillingWorker(parent_pid, size)],
     )
 
 
-def multitest_kills_worker():
+@testsuite
+class SimpleSuite(object):
+    @testcase
+    def test_simple(self, env, result):
+        pass
+
+
+def multitest_kills_worker(parent_pid):
     """To kill all child workers."""
-    os.kill(os.getpid(), 9)
+    if os.getpid() != parent_pid:  # Main process should not be killed
+        os.kill(os.getpid(), 9)
+    else:
+        return MultiTest(name="MTestKiller", suites=[SimpleSuite()])
 
 
 def schedule_tests_to_pool(plan, pool, schedule_path=None, **pool_cfg):
